DPC-236: Code Climate cleanups and Hibernate bundle refactors
Cleaned up a couple of code climate warnings.

Moved the Hibernate classes into DB specific packages.

Refactored the entity scanning into its own helper class.
