+++ dpc-aggregation/src/main/java/gov/cms/dpc/aggregation/DPCAggregationService.java
@@ -4,8 +4,8 @@ import ca.mestevens.java.configuration.bundle.TypesafeConfigurationBundle;
 import com.hubspot.dropwizard.guicier.GuiceBundle;
 import com.squarespace.jersey2.guice.JerseyGuiceUtils;
 import gov.cms.dpc.bluebutton.BlueButtonClientModule;
-import gov.cms.dpc.common.hibernate.DPCQueueHibernateBundle;
-import gov.cms.dpc.common.hibernate.DPCQueueHibernateModule;
+import gov.cms.dpc.common.hibernate.queue.DPCQueueHibernateBundle;
+import gov.cms.dpc.common.hibernate.queue.DPCQueueHibernateModule;
 import gov.cms.dpc.common.utils.EnvironmentParser;
 import gov.cms.dpc.queue.JobQueueModule;
 import io.dropwizard.Application;
