+++ dpc-api/src/main/java/gov/cms/dpc/api/DPCAPIConfiguration.java
@@ -2,9 +2,9 @@ package gov.cms.dpc.api;
 
 import ca.mestevens.java.configuration.TypesafeConfiguration;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import gov.cms.dpc.common.hibernate.IDPCAuthDatabase;
-import gov.cms.dpc.common.hibernate.IDPCDatabase;
-import gov.cms.dpc.common.hibernate.IDPCQueueDatabase;
+import gov.cms.dpc.common.hibernate.auth.IDPCAuthDatabase;
+import gov.cms.dpc.common.hibernate.attribution.IDPCDatabase;
+import gov.cms.dpc.common.hibernate.queue.IDPCQueueDatabase;
 import gov.cms.dpc.fhir.configuration.DPCFHIRConfiguration;
 import gov.cms.dpc.fhir.configuration.IDPCFHIRConfiguration;
 import io.dropwizard.client.JerseyClientConfiguration;
