+++ dpc-attribution/src/main/java/gov/cms/dpc/attribution/AttributionAppModule.java
@@ -8,8 +8,8 @@ import gov.cms.dpc.attribution.jdbi.*;
 import gov.cms.dpc.attribution.macaroons.BakeryProvider;
 import gov.cms.dpc.attribution.resources.v1.*;
 import gov.cms.dpc.attribution.tasks.TruncateDatabase;
-import gov.cms.dpc.common.hibernate.DPCHibernateBundle;
-import gov.cms.dpc.common.hibernate.DPCManagedSessionFactory;
+import gov.cms.dpc.common.hibernate.attribution.DPCHibernateBundle;
+import gov.cms.dpc.common.hibernate.attribution.DPCManagedSessionFactory;
 import gov.cms.dpc.macaroons.MacaroonBakery;
 import gov.cms.dpc.macaroons.store.hibernate.HibernateKeyStore;
 import gov.cms.dpc.macaroons.thirdparty.IThirdPartyKeyStore;
