+++ dpc-common/src/main/java/gov/cms/dpc/common/hibernate/attribution/DPCManagedSessionFactory.java
@@ -1,4 +1,4 @@
-package gov.cms.dpc.common.hibernate;
+package gov.cms.dpc.common.hibernate.attribution;
 
 import io.dropwizard.lifecycle.Managed;
 import org.hibernate.SessionFactory;
@@ -7,11 +7,11 @@ import org.hibernate.SessionFactory;
  * {@link Managed} interface that wraps a Hibernate {@link SessionFactory} and ensures that it is shutdown correctly when the service exits.
  * This is necessary because we manually inject the SessionFactory into various classes and thus we take ownership of its lifecycle.
  */
-public class DPCAuthManagedSessionFactory implements Managed {
+public class DPCManagedSessionFactory implements Managed {
 
     private final SessionFactory sessionFactory;
 
-    public DPCAuthManagedSessionFactory(SessionFactory factory) {
+    public DPCManagedSessionFactory(SessionFactory factory) {
         this.sessionFactory = factory;
     }
