+++ dpc-common/src/main/java/gov/cms/dpc/common/hibernate/attribution/IDPCDatabase.java
@@ -1,4 +1,4 @@
-package gov.cms.dpc.common.hibernate;
+package gov.cms.dpc.common.hibernate.attribution;
 
 import io.dropwizard.db.DataSourceFactory;
