+++ dpc-common/src/main/java/gov/cms/dpc/common/hibernate/auth/DPCAuthHibernateBundle.java
@@ -1,0 +1,39 @@
+package gov.cms.dpc.common.hibernate.auth;
+
+import com.google.inject.Inject;
+import io.dropwizard.Configuration;
+import io.dropwizard.ConfiguredBundle;
+import io.dropwizard.db.PooledDataSourceFactory;
+import io.dropwizard.hibernate.HibernateBundle;
+import io.dropwizard.hibernate.SessionFactoryFactory;
+
+import javax.inject.Singleton;
+
+import static gov.cms.dpc.common.hibernate.EntityScanner.applicationEntities;
+
+/**
+ * Custom Hibernate bundle, which allows us to inject the {@link org.hibernate.SessionFactory} into non-resource types.
+ * By default, the bundle scans the {@link DPCAuthHibernateBundle#PREFIX_STRING} for a list of annotated entities.
+ *
+ * @param <T> - Configuration class type
+ */
+@Singleton
+public class DPCAuthHibernateBundle<T extends Configuration & IDPCAuthDatabase> extends HibernateBundle<T> implements ConfiguredBundle<T> {
+
+    public static String PREFIX_STRING = "gov.cms.dpc.api.entities";
+
+    @Inject
+    public DPCAuthHibernateBundle() {
+        super(applicationEntities(PREFIX_STRING), new SessionFactoryFactory());
+    }
+
+    @Override
+    public PooledDataSourceFactory getDataSourceFactory(T configuration) {
+        return configuration.getAuthDatabase();
+    }
+
+    @Override
+    protected String name() {
+        return "hibernate.auth";
+    }
+}
