+++ dpc-common/src/main/java/gov/cms/dpc/common/hibernate/auth/IDPCAuthDatabase.java
@@ -1,4 +1,4 @@
-package gov.cms.dpc.common.hibernate;
+package gov.cms.dpc.common.hibernate.auth;
 
 import io.dropwizard.db.DataSourceFactory;
