+++ dpc-common/src/main/java/gov/cms/dpc/common/hibernate/queue/DPCQueueHibernateModule.java
@@ -1,4 +1,4 @@
-package gov.cms.dpc.common.hibernate;
+package gov.cms.dpc.common.hibernate.queue;
 
 import com.google.inject.Binder;
 import com.google.inject.Provides;
