+++ dpc-common/src/main/java/gov/cms/dpc/common/hibernate/queue/DPCQueueManagedSessionFactory.java
@@ -1,4 +1,4 @@
-package gov.cms.dpc.common.hibernate;
+package gov.cms.dpc.common.hibernate.queue;
 
 import io.dropwizard.lifecycle.Managed;
 import org.hibernate.SessionFactory;
