+++ dpc-queue/src/test/java/gov/cms/dpc/queue/QueueHealthTest.java
@@ -1,7 +1,7 @@
 package gov.cms.dpc.queue;
 
 import com.codahale.metrics.MetricRegistry;
-import gov.cms.dpc.common.hibernate.DPCQueueManagedSessionFactory;
+import gov.cms.dpc.common.hibernate.queue.DPCQueueManagedSessionFactory;
 import gov.cms.dpc.queue.exceptions.JobQueueUnhealthy;
 import gov.cms.dpc.queue.health.JobQueueHealthCheck;
 import org.hibernate.Session;
